+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- typo + conditional to orderform
+
 ## [0.12.8] - 2023-08-01
 
 ### Fixed